pyproject.toml
[tool.poetry]
name = "bitcoin-clarity"
version = "0.1.0"
description = ""
authors = ["Mfon Nwa <mfonnwa@gmail.com>"]
readme = "README.md"
packages = [{include = "bitcoin_clarity"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
